Actually use the CaloTruthAccumulator with Graph
